feat: conditional rendering of supplier balance added to header
